python 3.10 & Qt6.2 support, TreesAreMemory3 commit before refactoring

+++ kataja/PaletteManager.py
@@ -323,7 +323,7 @@ class PaletteManager:
         # polynomial approximation of how saturation and value are related to hue in Solarized
         saturation = 0.0000132 * h3 - 0.0068 * h2 + 0.7769 * hue + 77.188
         value = -7.07E-6 * h3 + 0.0042 * h2 - 0.6383 * hue + 88.615
-        return QColor().fromHsv(hue, min(255, saturation * 2.55), min(255, value * 2.55))
+        return QColor().fromHsv(hue, min(255, int(saturation * 2.55)), min(255, int(value * 2.55)))
 
     def is_custom(self):
         return self.theme_key in self.custom_themes
